+++ x/stake/handler.go
@@ -11,10 +11,10 @@ func NewHandler(k Keeper) sdk.Handler {
 	return func(ctx sdk.Context, msg sdk.Msg) sdk.Result {
 		// NOTE msg already has validate basic run
 		switch msg := msg.(type) {
-		case MsgDeclareCandidacy:
-			return handleMsgDeclareCandidacy(ctx, msg, k)
-		case MsgEditCandidacy:
-			return handleMsgEditCandidacy(ctx, msg, k)
+		case MsgCreateValidator:
+			return handleMsgCreateValidator(ctx, msg, k)
+		case MsgEditValidator:
+			return handleMsgEditValidator(ctx, msg, k)
 		case MsgDelegate:
 			return handleMsgDelegate(ctx, msg, k)
 		case MsgUnbond:
@@ -39,7 +39,7 @@ func NewEndBlocker(k Keeper) sdk.EndBlocker {
 // These functions assume everything has been authenticated,
 // now we just perform action and save
 
-func handleMsgDeclareCandidacy(ctx sdk.Context, msg MsgDeclareCandidacy, k Keeper) sdk.Result {
+func handleMsgCreateValidator(ctx sdk.Context, msg MsgCreateValidator, k Keeper) sdk.Result {
 
 	// check to see if the pubkey or sender has been registered before
 	_, found := k.GetValidator(ctx, msg.ValidatorAddr)
@@ -57,7 +57,7 @@ func handleMsgDeclareCandidacy(ctx sdk.Context, msg MsgDeclareCandidacy, k Keepe
 	k.setValidator(ctx, validator)
 	k.setValidatorByPubKeyIndex(ctx, validator)
 	tags := sdk.NewTags(
-		"action", []byte("declareCandidacy"),
+		"action", []byte("createValidator"),
 		"validator", msg.ValidatorAddr.Bytes(),
 		"moniker", []byte(msg.Description.Moniker),
 		"identity", []byte(msg.Description.Identity),
@@ -75,7 +75,7 @@ func handleMsgDeclareCandidacy(ctx sdk.Context, msg MsgDeclareCandidacy, k Keepe
 	}
 }
 
-func handleMsgEditCandidacy(ctx sdk.Context, msg MsgEditCandidacy, k Keeper) sdk.Result {
+func handleMsgEditValidator(ctx sdk.Context, msg MsgEditValidator, k Keeper) sdk.Result {
 
 	// validator must already be registered
 	validator, found := k.GetValidator(ctx, msg.ValidatorAddr)
@@ -95,7 +95,7 @@ func handleMsgEditCandidacy(ctx sdk.Context, msg MsgEditCandidacy, k Keeper) sdk
 
 	k.updateValidator(ctx, validator)
 	tags := sdk.NewTags(
-		"action", []byte("editCandidacy"),
+		"action", []byte("editValidator"),
 		"validator", msg.ValidatorAddr.Bytes(),
 		"moniker", []byte(msg.Description.Moniker),
 		"identity", []byte(msg.Description.Identity),
@@ -207,14 +207,14 @@ func handleMsgUnbond(ctx sdk.Context, msg MsgUnbond, k Keeper) sdk.Result {
 	bond.Shares = bond.Shares.Sub(delShares)
 
 	// remove the bond
-	revokeCandidacy := false
+	revokeValidator := false
 	if bond.Shares.IsZero() {
 
 		// if the bond is the owner of the validator then
-		// trigger a revoke candidacy
+		// trigger a revoke validator
 		if bytes.Equal(bond.DelegatorAddr, validator.Owner) &&
 			validator.Revoked == false {
-			revokeCandidacy = true
+			revokeValidator = true
 		}
 
 		k.removeDelegation(ctx, bond)
@@ -233,7 +233,7 @@ func handleMsgUnbond(ctx sdk.Context, msg MsgUnbond, k Keeper) sdk.Result {
 
 	/////////////////////////////////////
 	// revoke validator if necessary
-	if revokeCandidacy {
+	if revokeValidator {
 		validator.Revoked = true
 	}