+++ CHANGELOG.md
@@ -11,6 +11,8 @@ IMPROVEMENTS
 * auth module uses go-wire codec instead of 'encoding/json'
 
 FIXES
+* [cli] fixed cli-bash tests
+* [ci] added cli-bash tests
 
 ## 0.18.1
 
@@ -100,7 +102,7 @@ FEATURES
 * [gaiacli] Support queries for candidates, delegator-bonds
 * [gaiad] Added `gaiad export` command to export current state to JSON
 * [x/bank] Tx tags with sender/recipient for indexing & later retrieval
-* [x/stake] Tx tags with delegator/candidate for delegation & unbonding, and candidate info for declare candidate / edit candidacy
+* [x/stake] Tx tags with delegator/candidate for delegation & unbonding, and candidate info for declare candidate / edit validator
 
 IMPROVEMENTS
 
@@ -134,7 +136,7 @@ BREAKING CHANGES
 
 FEATURES:
 
-* Gaia stake commands include, DeclareCandidacy, EditCandidacy, Delegate, Unbond
+* Gaia stake commands include, CreateValidator, EditValidator, Delegate, Unbond
 * MountStoreWithDB without providing a custom store works.
 * Repo is now lint compliant / GoMetaLinter with tendermint-lint integrated into CI
 * Better key output, pubkey go-amino hex bytes now output by default